refactor(tests): move the pull-decision tests next to the decision
Adding them to the lifecycle suite put that file at 558 of its 500 code-line
limit. They pin what acquire_service_image decides, so images.rs is where they
belong — the same way prefetch.rs keeps its own.

No test changes: the three assertions and the shared libpod stand-in move
verbatim.

Refs #1218
